yund/requirements.txt
django==2.0.4
djangorestframework==3.11.0
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
six==1.10.0
whoosh==2.7.4
django-haystack==2.8.1
mysqlclient==1.4.4
coreapi==2.3.3
AgentLogin==3.0.0
Pillow==5.4.1
pyJWT==1.7.1